Remove old dependencies and add new ones
removed old frontend pages
renamed new pages
