+++ pages/login_page.py
@@ -4,38 +4,33 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webdriver import WebDriver
 
 from pages.base_page import BasePage, Locator
+from pages.nav_bar_page import NavBar
 
 logger = logging.getLogger(__name__)
 
 
-class ParabankLoginSection(BasePage):
-    USERNAME_FIELD: Locator = (By.XPATH, "//input[@name='username']", 5)
-    PASSWORD_FIELD: Locator = (By.XPATH, "//input[@name='password']", 5)
-    LOGIN_BUTTON: Locator = (By.XPATH, "//input[@value='Log In']", 5)
-
-    GREETING_MESSAGE: Locator = (By.XPATH, "//p[@class='smallText']", 5)
-    ERROR_TITLE: Locator = (By.XPATH, "//h1[normalize-space()='Error!']", 5)
-    ERROR_MESSAGE: Locator = (By.XPATH, "//p[@class='error']", 5)
-
-    error_message_empty_fields = "Please enter a username and password."
-    error_message_invalid_credentials = (
-        "The username and password could not be verified."
-    )
+class LoginPage(BasePage):
+    EMAIL_INPUT: Locator = (By.ID, "email", 5)
+    PASSWORD_INPUT: Locator = (By.ID, "password", 5)
+    LOGIN_BUTTON: Locator = (By.CSS_SELECTOR, "[data-test='login-submit']", 5)
 
     def __init__(self, driver: WebDriver):
         super().__init__(driver)
+        self.url = "https://practicesoftwaretesting.com/auth/login"
+        self.nav_bar = NavBar(self.driver)
 
-    def login(self, username: str, password: str) -> None:
-        self.click(self.USERNAME_FIELD)
-        self.input_text(self.USERNAME_FIELD, username)
-        self.click(self.PASSWORD_FIELD)
-        self.input_text(self.PASSWORD_FIELD, password)
+    def fill_email(self, email: str) -> None:
+        self.click(self.EMAIL_INPUT)
+        self.input_text(self.EMAIL_INPUT, email)
+
+    def fill_password(self, password: str) -> None:
+        self.click(self.PASSWORD_INPUT)
+        self.input_text(self.PASSWORD_INPUT, password)
+
+    def click_login(self) -> None:
         self.click(self.LOGIN_BUTTON)
 
-    def get_greeting_message(self) -> str:
-        result: str = self.get_text(self.GREETING_MESSAGE)
-        return result
-
-    def is_error_title_located(self) -> bool:
-        result: bool = self.is_element_located(self.ERROR_TITLE)
-        return result
+    def login(self, username: str, password: str) -> None:
+        self.fill_email(username)
+        self.fill_password(password)
+        self.click_login()
